requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.3
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
apache-airflow==3.1.7
apache-airflow-core==3.1.7
apache-airflow-providers-celery==3.15.2
apache-airflow-providers-common-compat==1.13.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.30.4
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-standard==1.11.0
apache-airflow-task-sdk==1.1.7
appdirs==1.4.4
appnope==0.1.4
arch==8.0.0
argcomplete==3.6.3
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
awscrt==0.29.2
babel==2.18.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
boto3==1.42.39
botocore==1.42.39
brotli==1.2.0
cachetools==6.2.6
cadwyn==5.4.6
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
cmdstanpy==1.3.0
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==44.0.3
cycler==0.12.1
Cython==3.2.4
databricks-sdk==0.85.0
debugpy==1.8.19
decorator==5.2.1
deepdiff==8.6.1
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
dill==0.4.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
dotenv==0.9.9
dynaconf==3.2.12
email-validator==2.3.0
evidently==0.4.33
executing==2.2.1
Faker==40.4.0
fastapi==0.117.1
fastapi-cli==0.0.20
fastjsonschema==2.21.2
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
flower==2.0.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gevent==25.9.1
geventhttpclient==2.3.7
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenback==1.3.0
greenlet==3.3.1
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
holidays==0.89
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
humanize==4.15.0
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
invoke==2.2.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
json5==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kagglehub==1.0.0
kagglesdk==0.1.15
kiwisolver==1.4.9
kombu==5.6.2
lazy-object-proxy==1.12.0
libcst==1.8.6
lightgbm==4.6.0
linkify-it-py==2.0.3
litestar==2.21.0
litestar-htmx==0.5.0
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
methodtools==0.4.7
mistune==3.2.0
ml_dtypes==0.5.4
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
mock==4.0.3
more-itertools==10.8.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.0
mypy_extensions==1.1.0
narwhals==2.15.0
natsort==8.4.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.24.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.7.0
orderly-set==5.5.0
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.5
pathspec==1.0.4
patsy==1.0.2
pendulum==3.2.0
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==5.24.1
pluggy==1.6.0
pmdarima==2.1.1
polyfactory==3.2.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.3.0
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.11.0
PyNaCl==1.6.2
pyparsing==3.3.2
pytest==9.0.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-rapidjson==1.23
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
PyYAML-ft==8.0.0
pyzmq==27.1.0
redis==6.4.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rich-argparse==1.7.2
rich-click==1.9.7
rich-toolkit==0.18.1
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.0
s3transfer==0.16.0
sagemaker==3.4.0
sagemaker-core==2.4.0
sagemaker-mlflow==0.2.0
sagemaker-mlops==1.4.0
sagemaker-schema-inference-artifacts==0.0.5
sagemaker-serve==1.4.0
sagemaker-train==1.4.0
schema==0.7.8
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setproctitle==1.3.7
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
skops==0.13.0
smdebug-rulesconfig==1.0.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
stack-data==0.6.3
stanio==0.5.1
starlette==0.48.0
statsmodels==0.14.6
streamlit==1.54.0
structlog==25.5.0
svcs==25.1.0
sympy==1.14.0
tabulate==0.9.0
tblib==3.2.2
tenacity==9.1.2
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
tritonclient==2.65.0
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uc-micro-py==1.0.3
ujson==5.11.0
universal_pathlib==0.2.6
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.1
watchtower==3.4.0
wcwidth==0.5.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.5
wirerope==1.0.0
wrapt==2.1.1
xgboost==3.1.3
yarl==1.22.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
